stboot: use OSImage interface
- change from bootconfig(jsonboot) to OSImage
- tune steps of bootflow
  - add dummy txt-detection step

Since OSImage serves as a common interface for the use of kexec, this commit
Introduces OSImage in stboot. Thus there is no more need for the bootconfig
struct (recently moved to pkg/boot/jsonboo).
Further this commit defines the different steps of the bootflow more precisely.
One new step is the TXT detection, but this is not yet implementet.

Signed-off-by: Jens Drenhaus <jens.drenhaus@9elements.com>